+++ requirements.txt
@@ -1,0 +1,52 @@
+aiohttp==3.8.4
+aiosignal==1.3.1
+async-timeout==4.0.2
+asyncio==3.4.3
+attrs==23.1.0
+blinker==1.6.2
+cachetools==5.3.2
+certifi==2023.11.17
+charset-normalizer==3.1.0
+click==8.1.3
+colorama==0.4.6
+Flask==2.3.2
+Flask-SQLAlchemy==3.0.3
+frozenlist==1.3.3
+google-api-core==2.15.0
+google-api-python-client==2.115.0
+google-auth==2.26.2
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.0
+googleapis-common-protos==1.62.0
+greenlet==2.0.2
+httplib2==0.22.0
+idna==3.4
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.2
+multidict==6.0.4
+mysql-connector==2.2.9
+mysql-connector-python==8.0.33
+nextcord==2.5.0
+numpy==1.24.3
+oauthlib==3.2.2
+pandas==2.0.1
+protobuf==4.25.2
+pyasn1==0.5.1
+pyasn1-modules==0.3.0
+pyparsing==3.1.1
+python-dateutil==2.8.2
+python-dotenv==1.0.1
+pytz==2023.3
+requests==2.31.0
+requests-oauthlib==1.3.1
+rsa==4.9
+simplejson==3.19.1
+six==1.16.0
+SQLAlchemy==2.0.14
+typing_extensions==4.5.0
+tzdata==2023.3
+uritemplate==4.1.1
+urllib3==2.1.0
+Werkzeug==2.3.4
+yarl==1.9.2
