+++ server/openai/chatbot.py
@@ -1,0 +1,4 @@
+import pathlib
+import textwrap
+
+# import google.generativeai as genai
